+++ quartz/src/commonMain/kotlin/com/vitorpamplona/quartz/nipBEBle/protocol/BleMessageChunker.kt
@@ -18,8 +18,9 @@
  * AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM, OUT OF OR IN CONNECTION
  * WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.
  */
-package com.vitorpamplona.quartz.nipBEBle
+package com.vitorpamplona.quartz.nipBEBle.protocol
 
+import com.vitorpamplona.quartz.nipBEBle.BleConfig
 import com.vitorpamplona.quartz.utils.Deflate
 
 /**
