+++ quartz/src/commonMain/kotlin/com/vitorpamplona/quartz/nipBEBle/transport/BleTransport.kt
@@ -18,7 +18,9 @@
  * AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM, OUT OF OR IN CONNECTION
  * WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.
  */
-package com.vitorpamplona.quartz.nipBEBle
+package com.vitorpamplona.quartz.nipBEBle.transport
+
+import com.vitorpamplona.quartz.nipBEBle.BlePeer
 
 /**
  * Platform-agnostic interface for BLE transport operations.
