+++ quartz/src/commonTest/kotlin/com/vitorpamplona/quartz/nipBEBle/protocol/BleMessageChunkerTest.kt
@@ -18,7 +18,7 @@
  * AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM, OUT OF OR IN CONNECTION
  * WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.
  */
-package com.vitorpamplona.quartz.nipBEBle
+package com.vitorpamplona.quartz.nipBEBle.protocol
 
 import kotlin.test.Test
 import kotlin.test.assertEquals
